[arcane,utils] Déplace 'ParallelLoopOptions' dans son propre fichier.
